MDD: Don't warn about failure to remove old streaming files if they don't exist

+++ mdd/mdd.pl
@@ -145,7 +145,9 @@ eval 'use MDD::XOSD';
 $no_xosd++ if $@;
 
 # Get rid of old log files in case root owns them
-unlink '/tmp/vlc.out' or $log->warn("Can't remove old /tmp/vlc.out: $!");
+if ( -e '/tmp/vlc.out') {
+    unlink '/tmp/vlc.out' or $log->warn("Can't remove old /tmp/vlc.out: $!");
+}
 
 my $user = $config{user} || 'mdd';
 # Change euid/uid
